Move xarray argument helpers to xarray.py (and tests accordingly)
