+++ src/metpy/calc/thermo.py
@@ -2,8 +2,6 @@
 # Distributed under the terms of the BSD 3-Clause License.
 # SPDX-License-Identifier: BSD-3-Clause
 """Contains a collection of thermodynamic calculations."""
-import functools
-from inspect import signature
 import warnings
 
 import numpy as np
@@ -18,43 +16,11 @@ from ..cbook import broadcast_indices
 from ..interpolate.one_dimension import interpolate_1d
 from ..package_tools import Exporter
 from ..units import check_units, concatenate, units
-from ..xarray import preprocess_and_wrap
+from ..xarray import add_vertical_dim_from_xarray, preprocess_and_wrap
 
 exporter = Exporter(globals())
 
 sat_pressure_0c = 6.112 * units.millibar
-
-
-def add_vertical_dim_from_xarray(func):
-    """Fill in optional vertical_dim from DataArray argument."""
-    @functools.wraps(func)
-    def wrapper(*args, **kwargs):
-        bound_args = signature(func).bind(*args, **kwargs)
-        bound_args.apply_defaults()
-
-        # Search for DataArray in arguments
-        dataarray_arguments = [
-            value for value in bound_args.arguments.values()
-            if isinstance(value, xr.DataArray)
-        ]
-
-        # Fill in vertical_dim
-        if (
-            len(dataarray_arguments) > 0
-            and 'vertical_dim' in bound_args.arguments
-        ):
-            try:
-                bound_args.arguments['vertical_dim'] = (
-                    dataarray_arguments[0].metpy.find_axis_number('vertical')
-                )
-            except AttributeError:
-                # If axis number not found, fall back to default but warn.
-                warnings.warn(
-                    'Vertical dimension number not found. Defaulting to initial dimension.'
-                )
-
-        return func(*bound_args.args, **bound_args.kwargs)
-    return wrapper
 
 
 @exporter.export
