+++ tests/test_xarray.py
@@ -12,7 +12,14 @@ from metpy.plots.mapping import CFProjection
 from metpy.testing import (assert_almost_equal, assert_array_almost_equal, assert_array_equal,
                            get_test_data, needs_cartopy)
 from metpy.units import DimensionalityError, units
-from metpy.xarray import check_axis, check_matching_coordinates, preprocess_and_wrap
+from metpy.xarray import (
+    add_grid_arguments_from_xarray,
+    add_vertical_dim_from_xarray,
+    check_axis,
+    check_matching_coordinates,
+    grid_deltas_from_dataarray,
+    preprocess_and_wrap
+)
 
 
 @pytest.fixture
@@ -1281,3 +1288,156 @@ def test_preprocess_and_wrap_with_variable():
     xr.testing.assert_identical(func(data1, data2), expected_12)
     assert isinstance(result_21, units.Quantity)
     assert_array_equal(func(data2, data1), expected_21)
+
+
+def test_grid_deltas_from_dataarray_lonlat(test_da_lonlat):
+    """Test grid_deltas_from_dataarray with a lonlat grid."""
+    dx, dy = grid_deltas_from_dataarray(test_da_lonlat)
+    true_dx = np.array([[[321609.59212064, 321609.59212065, 321609.59212064],
+                         [310320.85961483, 310320.85961483, 310320.85961483],
+                         [297980.72966733, 297980.72966733, 297980.72966733],
+                         [284629.6008561, 284629.6008561, 284629.6008561]]]) * units.m
+    true_dy = np.array([[[369603.78775948, 369603.78775948, 369603.78775948, 369603.78775948],
+                         [369802.28173967, 369802.28173967, 369802.28173967, 369802.28173967],
+                         [370009.56291098, 370009.56291098, 370009.56291098,
+                          370009.56291098]]]) * units.m
+    assert_array_almost_equal(dx, true_dx, 5)
+    assert_array_almost_equal(dy, true_dy, 5)
+
+
+def test_grid_deltas_from_dataarray_xy(test_da_xy):
+    """Test grid_deltas_from_dataarray with a xy grid."""
+    dx, dy = grid_deltas_from_dataarray(test_da_xy)
+    true_dx = np.array([[[[500] * 3]]]) * units('km')
+    true_dy = np.array([[[[500]] * 3]]) * units('km')
+    assert_array_almost_equal(dx, true_dx, 5)
+    assert_array_almost_equal(dy, true_dy, 5)
+
+
+def test_grid_deltas_from_dataarray_actual_xy(test_da_xy, ccrs):
+    """Test grid_deltas_from_dataarray with a xy grid and kind='actual'."""
+    # Construct lon/lat coordinates
+    y, x = xr.broadcast(*test_da_xy.metpy.coordinates('y', 'x'))
+    lonlat = (ccrs.Geodetic(test_da_xy.metpy.cartopy_globe)
+              .transform_points(test_da_xy.metpy.cartopy_crs, x.values, y.values))
+    lon = lonlat[..., 0]
+    lat = lonlat[..., 1]
+    test_da_xy = test_da_xy.assign_coords(
+        longitude=xr.DataArray(lon, dims=('y', 'x'), attrs={'units': 'degrees_east'}),
+        latitude=xr.DataArray(lat, dims=('y', 'x'), attrs={'units': 'degrees_north'}))
+
+    # Actually test calculation
+    dx, dy = grid_deltas_from_dataarray(test_da_xy, kind='actual')
+    true_dx = [[[[494426.3249766, 493977.6028005, 493044.0656467],
+                 [498740.2046073, 498474.9771064, 497891.6588559],
+                 [500276.2649627, 500256.3440237, 500139.9484845],
+                 [498740.6956936, 499045.0391707, 499542.7244501]]]] * units.m
+    true_dy = [[[[496862.4106337, 496685.4729999, 496132.0732114, 495137.8882404],
+                 [499774.9676486, 499706.3354977, 499467.5546773, 498965.2587818],
+                 [499750.8962991, 499826.2263137, 500004.4977747, 500150.9897759]]]] * units.m
+    assert_array_almost_equal(dx, true_dx, 3)
+    assert_array_almost_equal(dy, true_dy, 3)
+
+
+def test_grid_deltas_from_dataarray_nominal_lonlat(test_da_lonlat):
+    """Test grid_deltas_from_dataarray with a lonlat grid and kind='nominal'."""
+    dx, dy = grid_deltas_from_dataarray(test_da_lonlat, kind='nominal')
+    true_dx = [[[3.333333] * 3]] * units.degrees
+    true_dy = [[[3.333333]] * 3] * units.degrees
+    assert_array_almost_equal(dx, true_dx, 5)
+    assert_array_almost_equal(dy, true_dy, 5)
+
+
+@needs_cartopy
+def test_grid_deltas_from_dataarray_lonlat_assumed_order():
+    """Test grid_deltas_from_dataarray when dim order must be assumed."""
+    # Create test dataarray
+    lat, lon = np.meshgrid(np.array([38., 40., 42]), np.array([263., 265., 267.]))
+    test_da = xr.DataArray(
+        np.linspace(300, 250, 3 * 3).reshape((3, 3)),
+        name='temperature',
+        dims=('dim_0', 'dim_1'),
+        coords={
+            'lat': xr.DataArray(lat, dims=('dim_0', 'dim_1'),
+                                attrs={'units': 'degrees_north'}),
+            'lon': xr.DataArray(lon, dims=('dim_0', 'dim_1'), attrs={'units': 'degrees_east'})
+        },
+        attrs={'units': 'K'}).to_dataset().metpy.parse_cf('temperature')
+
+    # Run and check for warning
+    with pytest.warns(UserWarning, match=r'y and x dimensions unable to be identified.*'):
+        dx, dy = grid_deltas_from_dataarray(test_da)
+
+    # Check results
+    true_dx = [[222031.0111961, 222107.8492205],
+               [222031.0111961, 222107.8492205],
+               [222031.0111961, 222107.8492205]] * units.m
+    true_dy = [[175661.5413976, 170784.1311091, 165697.7563223],
+               [175661.5413976, 170784.1311091, 165697.7563223]] * units.m
+    assert_array_almost_equal(dx, true_dx, 5)
+    assert_array_almost_equal(dy, true_dy, 5)
+
+
+def test_grid_deltas_from_dataarray_invalid_kind(test_da_xy):
+    """Test grid_deltas_from_dataarray when kind is invalid."""
+    with pytest.raises(ValueError):
+        grid_deltas_from_dataarray(test_da_xy, kind='invalid')
+
+
+@needs_cartopy
+def test_add_grid_arguments_from_dataarray():
+    """Test the grid argument decorator for adding in arguments from xarray."""
+    @add_grid_arguments_from_xarray
+    def return_the_kwargs(
+        da,
+        dz=None,
+        dy=None,
+        dx=None,
+        vertical_dim=None,
+        y_dim=None,
+        x_dim=None,
+        latitude=None
+    ):
+        return {
+            'dz': dz,
+            'dy': dy,
+            'dx': dx,
+            'vertical_dim': vertical_dim,
+            'y_dim': y_dim,
+            'x_dim': x_dim,
+            'latitude': latitude
+        }
+
+    data = xr.DataArray(
+        np.zeros((1, 2, 2, 2)),
+        dims=('time', 'isobaric', 'lat', 'lon'),
+        coords={
+            'time': ['2020-01-01T00:00Z'],
+            'isobaric': (('isobaric',), [850., 700.], {'units': 'hPa'}),
+            'lat': (('lat',), [30., 40.], {'units': 'degrees_north'}),
+            'lon': (('lon',), [-100., -90.], {'units': 'degrees_east'})
+        }
+    ).to_dataset(name='zeros').metpy.parse_cf('zeros')
+    result = return_the_kwargs(data)
+    assert_array_almost_equal(result['dz'], [-150.] * units.hPa)
+    assert_array_almost_equal(result['dy'], [[[[1109415.632] * 2]]] * units.meter, 2)
+    assert_array_almost_equal(result['dx'], [[[[964555.967], [853490.014]]]] * units.meter, 2)
+    assert result['vertical_dim'] == 1
+    assert result['y_dim'] == 2
+    assert result['x_dim'] == 3
+    assert_array_almost_equal(
+        result['latitude'].metpy.unit_array,
+        [30., 40.] * units.degrees_north
+    )
+    # Verify latitude is xarray so can be broadcast,
+    # see https://github.com/Unidata/MetPy/pull/1490#discussion_r483198245
+    assert isinstance(result['latitude'], xr.DataArray)
+
+
+def test_add_vertical_dim_from_xarray():
+    """Test decorator for automatically determining the vertical dimension number."""
+    @add_vertical_dim_from_xarray
+    def return_vertical_dim(data, vertical_dim=None):
+        return vertical_dim
+    test_da = xr.DataArray(np.zeros((2, 2, 2, 2)), dims=('time', 'isobaric', 'y', 'x'))
+    assert return_vertical_dim(test_da) == 1
